Move aot inductor utils from test to Inductor for reusing
[ghstack-poisoned]
