+++ test/inductor/test_aot_inductor.py
@@ -12,6 +12,7 @@ import torch._inductor
 from torch._dynamo.testing import rand_strided, same
 from torch._dynamo.utils import counters
 from torch._inductor import config
+from torch._inductor.aot_inductor_utils import AOTIRunnerUtil
 from torch._inductor.exc import CppWrapperCodeGenError
 from torch._inductor.test_case import TestCase
 from torch._inductor.utils import cache_dir
@@ -55,11 +56,9 @@ if IS_WINDOWS and IS_CI:
 
 try:
     try:
-        from .test_aot_inductor_utils import AOTIRunnerUtil
         from .test_control_flow import CondModels, prepend_predicates
         from .test_torchinductor import copy_tests, requires_multigpu, TestFailure
     except ImportError:
-        from test_aot_inductor_utils import AOTIRunnerUtil
         from test_control_flow import CondModels, prepend_predicates
         from test_torchinductor import copy_tests, requires_multigpu, TestFailure
 except (unittest.SkipTest, ImportError) as e:
